[Bronze V] Title: 이상한 기호, Time: 168 ms, Memory: 17732 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args){
 		Scanner scan = new Scanner(System.in);
-		int a = scan.nextInt();
-		int b = scan.nextInt();
+		long a = scan.nextLong();
+		long b = scan.nextLong();
 		System.out.println((a+b)*(a-b));
 		}
